% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockdown.R
\name{lockdown}
\alias{lockdown}
\alias{lockdown_file}
\alias{lockdown_variable}
\title{Lock down a file or variable}
\usage{
lockdown_file(fp, cache_fp = ".lockdown_files.rds")

lockdown_variable(x, cache_fp = ".lockdown_vars.rds")
}
\arguments{
\item{fp}{The file path.}

\item{x}{The variable.}
}
\value{
Invisibly, returns \code{NA} if the file was not found in the cache,
  and \code{TRUE} if the file matches its value in the cache.  Will produce
  a VERY LOUD ERROR MESSAGE if the file was found in the cache and it does
  not match.
}
\description{
Lock down a file or variable
}
