% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockdown.R
\name{lockdownremove}
\alias{lockdownremove}
\alias{lockdown_remove_file}
\alias{lockdown_remove_variable}
\title{Remove a file or variable from the lockdown cache}
\usage{
lockdown_remove_file(fp, cache_fp = ".lockdown_files.rds")

lockdown_remove_variable(x, cache_fp = ".lockdown_vars.rds")
}
\arguments{
\item{fp}{The file path.}

\item{x}{The variable.}
}
\value{
Invisibly, returns \code{NA} if the file was not found in the cache,
  and \code{TRUE} if the file was found and removed.
}
\description{
Remove a file or variable from the lockdown cache
}
